finadvisor-ai/backend/pyproject.toml
[project]
name = "finadvisor-backend"
version = "1.0.0"
description = "FinAdvisor AI — Advanced Financial Advisory Agent Backend"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-groq>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langsmith>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "tiktoken>=0.8.0",
    "supabase>=2.9.0",
    "vecs>=0.4.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.6.0",
    "slowapi>=0.1.9",
    "yfinance>=0.2.48",
    "httpx>=0.28.0",
    "requests>=2.32.0",
    "tenacity>=9.0.0",
    "jinja2>=3.1.0",
    "structlog>=24.4.0",
    "mcp[cli]>=1.0.0",
    "apscheduler>=3.10.0",
    "pydantic[email]>=2.12.5",
    "bcrypt==4.0.1",
    "stripe>=5.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["finadvisor_backend"]
